requirements.txt
flask
requests
gunicorn
google_crc32c
google-cloud-secret-manager